requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
certifi==2025.7.14
click==8.2.1
dnspython==2.7.0
email-validator==2.2.0
fastapi==0.116.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
mako==1.3.10
markupsafe==3.0.2
packaging==25.0
pluggy==1.6.0
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pygments==2.19.2
pytest==8.4.1
pytest-asyncio==1.1.0
python-dotenv==1.1.1
ruff==0.12.4
sniffio==1.3.1
sqlalchemy==2.0.41
starlette==0.47.1
typing-extensions==4.14.1
typing-inspection==0.4.1
uvicorn==0.35.0
